refactor!: add plugin system
- externalized class validator adapter within its plugin
- removed the concept of extending the metadata store
- add BaseMetadataAdapter

BREAKING CHANGE: Extending class-fixtures-factory requires extending BaseMetadataAdapter instead of BaseMetadataStore
