+++ src/plugins/class-validator/index.ts
@@ -1,0 +1,4 @@
+import { MetadataStore } from '../../metadata';
+import { ClassValidatorAdapter } from './ClassValidatorAdapter';
+
+MetadataStore.setAdapter(new ClassValidatorAdapter());
